updated MChestKernel
